Moved AuditLog feature into core.audit package
