+++ build.gradle
@@ -114,6 +114,7 @@ javadoc {
     failOnError = false
     options.memberLevel = JavadocMemberLevel.PUBLIC
     options.author()
+    options.encoding = 'UTF-8'
 
     //### excludes ###
     //jda internals
